requirements.txt
streamlit
faiss-cpu
pypdf
python-docx
docx2txt
python-dotenv
sentence-transformers
openai
gTTS
bcrypt
torchvision
